user authentication

+++ controllers/expenses.js
@@ -2,7 +2,10 @@ const Expenses = require("../models/expenses");
 
 exports.addExpense = async (req, res, next) => {
   try {
-    const newExpense = await Expenses.create({ ...req.body });
+    const newExpense = await Expenses.create({
+      ...req.body,
+      userId: req.user.id,
+    });
     return res.status(200).json({ success: true, message: "expense created" });
   } catch (err) {
     return res
@@ -12,11 +15,12 @@ exports.addExpense = async (req, res, next) => {
 };
 
 exports.getExpenses = async (req, res, next) => {
-  await Expenses.findAll()
+  await Expenses.findAll({ where: { userId: req.user.id } })
     .then((expense) => {
       return res.status(200).json(expense);
     })
     .catch((err) => {
+      console.log(err);
       return res
         .status(500)
         .json({ success: false, message: "sometihng went wrong" });
@@ -25,6 +29,7 @@ exports.getExpenses = async (req, res, next) => {
 
 exports.deleteExpense = async (req, res, next) => {
   const expenseId = req.params.id;
-  await Expenses.destroy({ where: { id: expenseId } });
+  const userId = req.params.userId;
+  await Expenses.destroy({ where: { id: expenseId, userId: userId } });
   res.status(200).json({ success: true, message: "expense deleted" });
 };
